refactor: move controller types to shared dir
